Update the way `sh:targetClass` of box.getLabel() is rendered.
The orginal comment stated `use the sh:targetClass if present,
otherwise use the URI of the NodeShape`

This commit changes the behaviour of the method that it matches the
comment. Or at least the way I could interpret it.

We could also consider adding this feature as a CLI option.